requirements.txt
aiovk
aiosqlite
python-telegram-bot
